add) Home 화면 및 기본 Navigation 추가 작업
1) HomeScreen.kt, HomeScreenState.kt, HomeScreenViewModel.kt
MVI 패턴으로 State, Event, Effect 적용 버전으로 추가.

2) BaseViewModel.kt
모든 viewModel에 MVI 기본 패턴 틀을 위한 Base 추가

3) BookListApi.kt
중복 쿼리 제거.

+++ domain/book/src/main/java/lee/project/domain/book/model/MyBookModel.kt
@@ -8,5 +8,5 @@ data class MyBookModel(
     val link: String?, // 상품 링크 Url
     val myReview: String?, // 독후감
     val period: String, // 독서 기간
-    val rating: Int // 내 평점
+    val rating: Float // 내 평점
 )
